Tidy up a few loose ends and implement duplicate.
