+++ extendr-api/src/lib.rs
@@ -266,7 +266,7 @@ pub use robj_ndarray::*;
 pub use extendr_macros::*;
 //////////////////////////////////////////////////
 
-pub struct Cplx(f64, f64);
+pub struct Cplx(pub f64, pub f64);
 
 /// TRUE value eg. `r!(TRUE)`
 pub const TRUE: Bool = Bool(1);
